requirements.txt
dash
dash_core_components
dash_html_components
plotly.graph_objects
yfinance
pandas
numpy
pandas_datareader
